terminapp/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
